refine http-model ergonomics and move typed headers
