+++ endpoint/shared/src/test/scala-3/zio/blocks/endpoint/EndpointSpec.scala
@@ -330,7 +330,7 @@ object EndpointSpec extends ZIOSpecDefault {
           h == HttpCodec.Header[CodecKind.Request, String]("Authorization", Schema.string),
           b == HttpCodec.Body[CodecKind.Response, String](Schema.string),
           s == HttpCodec.StatusCodec(Some(Status.Created)),
-          a.isInstanceOf[HttpCodec[CodecKind.Request, zio.http.headers.Authorization.Bearer]]
+          a.isInstanceOf[HttpCodec[CodecKind.Request, zio.http.Header.Authorization.Bearer]]
         )
       },
       test("query codec") {
@@ -392,7 +392,7 @@ object EndpointSpec extends ZIOSpecDefault {
       },
       test("typed authorization schema failures surface SchemaError") {
         val codec = HttpCodec.bearerAuth.asInstanceOf[
-          HttpCodec.Header[CodecKind.Request, zio.http.headers.Authorization.Bearer]
+          HttpCodec.Header[CodecKind.Request, zio.http.Header.Authorization.Bearer]
         ]
         val result = codec.schema.fromDynamicValue(Schema.string.toDynamicValue("Basic Zm9vOmJhcg=="))
 
@@ -435,7 +435,7 @@ object EndpointSpec extends ZIOSpecDefault {
           typeCheck("""
             import zio.blocks.endpoint._
             import zio.http.Status
-            import zio.http.headers
+            import zio.http.Header
 
             val endpoint = Endpoint(Method.GET / "users")
               .auth(AuthType.Bearer)
@@ -443,7 +443,7 @@ object EndpointSpec extends ZIOSpecDefault {
 
             val auth = endpoint.auth
             val status = auth.unauthorizedStatus
-            val codec: HttpCodec[CodecKind.Request, headers.Authorization.Bearer] = auth.codec
+            val codec: HttpCodec[CodecKind.Request, Header.Authorization.Bearer] = auth.codec
           """)
         )(isRight)
       }
