--- http-model/shared/src/main/scala/zio/http/headers/CookieHeaders.scala
@@ -1,45 +1,0 @@
-/*
- * Copyright 2024-2026 John A. De Goes and the ZIO Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package zio.http.headers
-
-import zio.http._
-
-// --- 1. CookieHeader ---
-
-final case class CookieHeader(value: String) extends Header {
-  def headerName: String    = CookieHeader.name
-  def renderedValue: String = value
-}
-
-object CookieHeader extends Header.Typed[CookieHeader] {
-  val name: String                                       = "cookie"
-  def parse(value: String): Either[String, CookieHeader] = Right(CookieHeader(value))
-  def render(h: CookieHeader): String                    = h.renderedValue
-}
-
-// --- 2. SetCookieHeader ---
-
-final case class SetCookieHeader(value: String) extends Header {
-  def headerName: String    = SetCookieHeader.name
-  def renderedValue: String = value
-}
-
-object SetCookieHeader extends Header.Typed[SetCookieHeader] {
-  val name: String                                          = "set-cookie"
-  def parse(value: String): Either[String, SetCookieHeader] = Right(SetCookieHeader(value))
-  def render(h: SetCookieHeader): String                    = h.renderedValue
-}
