+++ http-model/shared/src/test/scala/zio/http/RequestSpec.scala
@@ -16,7 +16,8 @@
 
 package zio.http
 
-import zio.test._
+import _root_.zio.test._
+import zio.blocks.chunk.Chunk
 
 object RequestSpec extends HttpModelBaseSpec {
   def spec: Spec[TestEnvironment, Any] = suite("Request")(
@@ -75,7 +76,7 @@ object RequestSpec extends HttpModelBaseSpec {
       test("returns typed header from headers") {
         val headers = Headers("host" -> "example.com:8080")
         val request = Request(Method.GET, URL.fromPath(Path.root), headers, Body.empty, Version.`HTTP/1.1`)
-        val host    = request.header(zio.http.headers.Host)
+        val host    = request.header(Header.Host)
         assertTrue(
           host.isDefined,
           host.get.host == "example.com",
@@ -84,7 +85,7 @@ object RequestSpec extends HttpModelBaseSpec {
       },
       test("returns None for missing header") {
         val request = Request.get(URL.fromPath(Path.root))
-        assertTrue(request.header(zio.http.headers.Host).isEmpty)
+        assertTrue(request.header(Header.Host).isEmpty)
       }
     ),
     suite("contentType")(
@@ -178,6 +179,10 @@ object RequestSpec extends HttpModelBaseSpec {
       test("adds a header to request") {
         val request = Request.get(URL.fromPath(Path.root)).addHeader("Accept", "text/html")
         assertTrue(request.headers.rawGet("accept") == Some("text/html"))
+      },
+      test("adds a typed header to request") {
+        val request = Request.get(URL.fromPath(Path.root)).addHeader(Header.Host("example.com", Some(8080)))
+        assertTrue(request.header(Header.Host) == Some(Header.Host("example.com", Some(8080))))
       }
     ),
     suite("addHeaders")(
@@ -206,6 +211,33 @@ object RequestSpec extends HttpModelBaseSpec {
           .addHeader("Accept", "text/html")
           .setHeader("Accept", "application/json")
         assertTrue(request.headers.rawGet("accept") == Some("application/json"))
+      },
+      test("sets a typed header on request") {
+        val request = Request
+          .get(URL.fromPath(Path.root))
+          .setHeader(Header.Host("example.com", Some(8080)))
+        assertTrue(request.header(Header.Host) == Some(Header.Host("example.com", Some(8080))))
+      }
+    ),
+    suite("cookies")(
+      test("parses cookies from Cookie header") {
+        val request = Request.get(URL.fromPath(Path.root)).setHeader("Cookie", "session=abc123; theme=dark")
+        assertTrue(
+          request.cookies == Chunk(RequestCookie("session", "abc123"), RequestCookie("theme", "dark"))
+        )
+      },
+      test("returns empty cookies when Cookie header is absent") {
+        assertTrue(Request.get(URL.fromPath(Path.root)).cookies == Chunk.empty)
+      },
+      test("addCookie appends to existing Cookie header") {
+        val request = Request
+          .get(URL.fromPath(Path.root))
+          .addCookie(RequestCookie("session", "abc123"))
+          .addCookie(RequestCookie("theme", "dark"))
+        assertTrue(
+          request.headers.rawGet("cookie") == Some("session=abc123; theme=dark"),
+          request.cookies == Chunk(RequestCookie("session", "abc123"), RequestCookie("theme", "dark"))
+        )
       }
     ),
     suite("body (setter)")(
