+++ http-model/shared/src/test/scala/zio/http/ResponseSpec.scala
@@ -16,7 +16,8 @@
 
 package zio.http
 
-import zio.test._
+import _root_.zio.test._
+import zio.blocks.chunk.Chunk
 
 object ResponseSpec extends HttpModelBaseSpec {
   def spec: Spec[TestEnvironment, Any] = suite("Response")(
@@ -84,14 +85,14 @@ object ResponseSpec extends HttpModelBaseSpec {
       test("returns typed header from headers") {
         val headers  = Headers("content-length" -> "42")
         val response = Response(Status.Ok, headers, Body.empty, Version.`HTTP/1.1`)
-        val cl       = response.header(zio.http.headers.ContentLength)
+        val cl       = response.header(Header.ContentLength)
         assertTrue(
           cl.isDefined,
           cl.get.length == 42L
         )
       },
       test("returns None for missing header") {
-        assertTrue(Response.ok.header(zio.http.headers.ContentLength).isEmpty)
+        assertTrue(Response.ok.header(Header.ContentLength).isEmpty)
       }
     ),
     suite("contentType")(
@@ -142,6 +143,10 @@ object ResponseSpec extends HttpModelBaseSpec {
           response.headers.rawGet("content-type") == Some("text/plain; charset=UTF-8"),
           response.contentType == Some(response.body.contentType)
         )
+      },
+      test("accepts explicit status") {
+        val response = Response.text(Status.BadRequest, "hello world")
+        assertTrue(response.status == Status.BadRequest)
       }
     ),
     suite("Response.json")(
@@ -151,6 +156,21 @@ object ResponseSpec extends HttpModelBaseSpec {
           response.status == Status.Ok,
           response.headers.rawGet("content-type") == Some("application/json"),
           response.contentType == Some(ContentType.`application/json`)
+        )
+      },
+      test("accepts explicit status") {
+        val response = Response.json(Status.BadRequest, "{\"key\": \"value\"}")
+        assertTrue(response.status == Status.BadRequest)
+      }
+    ),
+    suite("Response.ok(body)")(
+      test("creates ok response with synced content-type header") {
+        val body     = Body.fromString("hello")
+        val response = Response.ok(body)
+        assertTrue(
+          response.status == Status.Ok,
+          response.body == body,
+          response.headers.rawGet("content-type") == Some(body.contentType.render)
         )
       }
     ),
@@ -183,6 +203,10 @@ object ResponseSpec extends HttpModelBaseSpec {
       test("adds a header") {
         val response = Response.ok.addHeader("X-Custom", "value")
         assertTrue(response.headers.rawGet("x-custom") == Some("value"))
+      },
+      test("adds a typed header") {
+        val response = Response.ok.addHeader(Header.ContentLength(42L))
+        assertTrue(response.header(Header.ContentLength) == Some(Header.ContentLength(42L)))
       }
     ),
     suite("Response addHeaders")(
@@ -207,6 +231,12 @@ object ResponseSpec extends HttpModelBaseSpec {
           .addHeader("X-Custom", "old")
           .setHeader("X-Custom", "new")
         assertTrue(response.headers.rawGet("x-custom") == Some("new"))
+      },
+      test("sets a typed header replacing existing") {
+        val response = Response.ok
+          .addHeader(Header.ContentLength(1L))
+          .setHeader(Header.ContentLength(42L))
+        assertTrue(response.header(Header.ContentLength) == Some(Header.ContentLength(42L)))
       }
     ),
     suite("Response body (setter)")(
@@ -247,6 +277,20 @@ object ResponseSpec extends HttpModelBaseSpec {
         val cookie   = ResponseCookie("session", "abc123")
         val response = Response.ok.addCookie(cookie)
         assertTrue(response.headers.has("set-cookie"))
+      }
+    ),
+    suite("Response cookies")(
+      test("parses all response cookies from Set-Cookie headers") {
+        val response = Response.ok
+          .addCookie(ResponseCookie("session", "abc123"))
+          .addCookie(ResponseCookie("theme", "dark"))
+        assertTrue(
+          response.cookies == Chunk(ResponseCookie("session", "abc123"), ResponseCookie("theme", "dark"))
+        )
+      },
+      test("skips invalid Set-Cookie headers") {
+        val response = Response.ok.addHeader("set-cookie", "invalid")
+        assertTrue(response.cookies == Chunk.empty)
       }
     )
   )
