+++ http-model/shared/src/test/scala/zio/http/headers/AuthHeadersSpec.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package zio.http.headers
+package zio.http
 
-import zio.test._
+import _root_.zio.test._
+import Header._
 
 object AuthHeadersSpec extends ZIOSpecDefault {
   def spec: Spec[TestEnvironment, Any] = suite("AuthHeaders")(
