+++ http-model/shared/src/test/scala/zio/http/headers/CorsHeadersSpec.scala
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package zio.http.headers
+package zio.http
 
-import zio.test._
+import _root_.zio.test._
 import zio.blocks.chunk.Chunk
 import zio.http.Method
+import Header._
 
 object CorsHeadersSpec extends ZIOSpecDefault {
   def spec: Spec[TestEnvironment, Any] = suite("CorsHeaders")(
